Initial Game Concept
Game concept: The room expands and contracts as time remaining increases
or decreases.  There will be collectibles throughout the room that will
make the timer increase for x seconds rather than decreasing for that
duration.  The only thing in the game that does not scale with time is
the player.  Since the max speed the player moves is fixed in world
space and scale, as the room gets bigger it takes longer to traverse the
level to collect bonus time and as it gets smaller it gets easier.