Add Collectible visual & enforce y = 0 for floor & min percent
Added a temporary material for the collectible objects
Updated the Box prefab so that the floor of the room is always at y = 0
to help with player physics not falling through the floor
Updated the Game Manager to have a minimum percent and for scaling down
to stop at that percent.  This is to prevent the player from either
always hitting a collectible or being thrown from the map as the room
gets smaller than the player.  Some kind of trigger will need to be
added to end the game when the time runs out.
The collectible and player are now on layers so the collectible cannot
be triggered by the environment.